refactor: rename inputItems directory
