+++ components/InputBar/InputItems/amount.js
