+++ components/InputBar/InputItems/content.js
